This should be a working version with the AuthZ server deployed. I cleaned up a bunch of files that were generating compiler warnings. Changed use of GrantedAuthorithyImpl to SimpleGrantedAuthority, as GrantedAuthorityImpl is deprecated and SGA is its replacement.
Added stub for public OAuth2AccessToken readAccessToken(String accessToken) to DefaultOAuth2ProviderTokenService.java. This method IS in the superclass and should be implemented, but was not before and for some reason the compiler only started complaining about it now, on my machine.

Moved @Autowired for ClientDetailsEntityService in OAuthConfirmationController.java from setter method to field declaration, as it was failing to deploy on my machine in the other position. If others have issues with this change please let me know.