We created an utility class called “UrlUtils”, which has a method for url normalization that uses the webfinger normalizer.
Modifications in classes :

- PlainAuthRequestUrlBuilder : We added a boolean “useNonce” to allow
the integration with Google's IdP that does not allow the nonce value
into request , the OpenID Connect protocol specifies that the nonce
value is optional.

- OIDCAuthenticationFilter : We added a boolean “useNonce”, during the
valition process the framework was waiting for a nonce value, however
its use is optional and some IdPs may not include it in their response;
We have also used the aforementioned "UrlUtils" class to normalize the
issuer value returned by the IdP at the time of validating the
authentication ticket.

- DynamicServerConfigurationService: Now uses the "UrlUtils" class to
normalize the url returned 
by the IdP during the end-points discovery process and then configure
the issuer with the normalized url.