Remove transitive commons-logging dependencies
We use slf4j instead; jcl-over-slf4j needs to be a compile-time dependency because we use it in several classes.

Unfortunately Maven does not have a way to exclude commons-logging globally, so we need to figure out which dependencies include it through `mvn dependency:tree` and specify an `<exclusion>` for each of them.

Finally, we upgrade to slf4j 1.7.12: http://www.slf4j.org/news.html.