Make apiAddClient in the client api return a HttpStatus.Conflict if you try to create a client with a used client id.
This fixes a bug where if you try to create a client with a client id that is already in use, you get an empty error message. Instead, now you get a message that tells you that the client couldn't be created because the client id is already in use.