Update TestDefaultIntrospectionResultAssembler.java
Hi Justin, I am trying to throw a ParseException where the actual value and expected value are not equal for the expiration parameters. I am trying to test the catch block. I have used the same parameters which are already used in other test cases in the same file. Can you please guide me on how to approach catch case in this particular scenario, I have already tried but I am not sure this is the right way. Thank you